20200422.1315 pipe_play not working
Got thread_play a server and got it working with a pipe. server always used "b" connection, client always used "a". Probably could have server use b to recv and a respond, while client uses a to request and expected response on b. Best way depends on having meaningfull names for the 2 connections.